requirements.txt
langchain
langchain-community
langchain-nvidia-ai-endpoints
python-dotenv
serpapi
beautifulsoup4
requests